Accessibility changes for "Month" view
Changes to angular-calendar for better accessibility as the "American Disability Act" becomes mainstream and other developers would have to comply with it too.
Current changes are for "Month" view as a screen reader (i.e. NVDA, TalkBack, and VoiceOver) would speak the number of rows and columns of the calendar grid.